Move imglyb_config to imglyb.config
