+++ setup.py
@@ -1,7 +1,7 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 version_info={}
-with open('imglyb_config.py', 'r') as f:
+with open('imglyb/config.py', 'r') as f:
     exec(f.read(), version_info)
 version = version_info['version']
 
@@ -12,7 +12,6 @@ setup(
     author='Philipp Hanslovsky, Curtis Rueden, Edward Evans',
     author_email='ctrueden@wisc.edu',
     description='A python module to bring together the worlds of numpy (Python) and ImgLib2 (Java).',
-    packages=['imglyb'],
-    py_modules=['imglyb_config'],
+    packages=find_packages(),
     install_requires=['numpy', 'jpype1', 'scyjava'],
 )
